backend/requirements.txt
amqp==5.3.1
aniso8601==9.0.1
bcrypt==4.2.1
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comtypes==1.4.8
decorator==5.1.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Dance==7.1.0
Flask-Logging==0.1.3
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SocketIO==5.5.0
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
git-filter-repo==2.47.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==3.0.2
moviepy==2.1.1
numpy==2.2.1
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.10
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pyttsx3==2.98
pytz==2024.2
pywin32==308
requests==2.32.3
requests-oauthlib==2.0.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
URLObject==2.4.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
